[#80] Add CPP to use `nominalDiffTimeToSeconds` if `time`'s version suffices
`nominalDiffTimeToSeconds` was introduced in `time-1.9.1` and has type
`nominalDiffTimeToSeconds :: NominalDiffTime -> Pico`, which is what
is needed for `toUrlPiece` in `ToHttpApiData`, now that after issue #77
there are instances for types from `Data.Fixed`.

As such, if the `time` library's version is high enough, that
function will be used. Otherwise,
`(realToFrac :: NominalDiffTime -> Pico)` will be used instead.